Move App to src
