+++ include/cron_sync.php
@@ -40,6 +40,11 @@
 
  */
 
+//Startup.
+require_once 'boot.php';
+
+use Friendica\Directory\App;
+
 // Debug stuff.
 ini_set('display_errors', 1);
 ini_set('log_errors', '0');
@@ -47,8 +52,6 @@ error_reporting(E_ALL ^ E_NOTICE);
 
 $start_syncing = time();
 
-//Startup.
-require_once 'boot.php';
 $a = new App;
 
 //Create a simple log function for CLI use.
