+++ index.php
@@ -1,7 +1,8 @@
 <?php
 
-require_once 'vendor/autoload.php';
 require_once 'boot.php';
+
+use Friendica\Directory\App;
 
 $a = new App;
